# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/checkcreneau": {
            "post": {
                "summary": "Vérifier creneau",
                "operationId": "2a09452c4289d37318b9cb98955ebe2f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_times": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-09-05 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/service": {
            "get": {
                "summary": "liste services",
                "operationId": "047755c8af84519fc8294b1fbfdb5c32",
                "parameters": [
                    {
                        "name": "idservicecateg",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}